+++ gtmfe/include/data.h
@@ -24,44 +24,44 @@
 
 #include "constants.h"
 
-extern float poppen[5];
-extern float maxpen;
-extern float eparam[11];
-extern float multConst[3]; /* for multiloop penalties. */
-extern float dangle[4][4][4][2]; /* Contain dangling energy values */
-extern float inter[31]; /* Contains size penalty for internal loops */
-extern float bulge[31]; /* Contain the size penalty for bulges */
-extern float hairpin[31]; /* Contains the size penalty for hairpin loops */
-extern float stack[256]; /* Stacking energy used to calculate energy of stack loops */
-extern float tstkh[256]; /* Terminal mismatch energy used in the calculations of hairpin loops */
-extern float tstki[256]; /* Terminal mismatch energy used in the calculations of internal loops */
-extern float tloop[maxtloop + 1][2];
+extern long double poppen[5];
+extern long double maxpen;
+extern long double eparam[11];
+extern long double multConst[3]; /* for multiloop penalties. */
+extern long double dangle[4][4][4][2]; /* Contain dangling energy values */
+extern long double inter[31]; /* Contains size penalty for internal loops */
+extern long double bulge[31]; /* Contain the size penalty for bulges */
+extern long double hairpin[31]; /* Contains the size penalty for hairpin loops */
+extern long double stack[256]; /* Stacking energy used to calculate energy of stack loops */
+extern long double tstkh[256]; /* Terminal mismatch energy used in the calculations of hairpin loops */
+extern long double tstki[256]; /* Terminal mismatch energy used in the calculations of internal loops */
+extern long double tloop[maxtloop + 1][2];
 extern int numoftloops;
-extern float iloop22[5][5][5][5][5][5][5][5]; /* 2*2 internal looops */
-extern float iloop21[5][5][5][5][5][5][5]; /* 2*1 internal loops */
-extern float iloop11[5][5][5][5][5][5]; /* 1*1 internal loops */
-extern float coax[6][6][6][6];
-extern float tstackcoax[6][6][6][6];
-extern float coaxstack[6][6][6][6];
-extern float tstack[6][6][6][6];
-extern float tstkm[6][6][6][6];
-extern float auend;
-extern float gubonus;
-extern float cint; /* cint, cslope, c3 are used for poly C hairpin loops */
-extern float cslope;
-extern float c3;
-extern float efn2a;
-extern float efn2b;
-extern float efn2c;
-extern float triloop[maxtloop + 1][2];
+extern long double iloop22[5][5][5][5][5][5][5][5]; /* 2*2 internal looops */
+extern long double iloop21[5][5][5][5][5][5][5]; /* 2*1 internal loops */
+extern long double iloop11[5][5][5][5][5][5]; /* 1*1 internal loops */
+extern long double coax[6][6][6][6];
+extern long double tstackcoax[6][6][6][6];
+extern long double coaxstack[6][6][6][6];
+extern long double tstack[6][6][6][6];
+extern long double tstkm[6][6][6][6];
+extern long double auend;
+extern long double gubonus;
+extern long double cint; /* cint, cslope, c3 are used for poly C hairpin loops */
+extern long double cslope;
+extern long double c3;
+extern long double efn2a;
+extern long double efn2b;
+extern long double efn2c;
+extern long double triloop[maxtloop + 1][2];
 extern int numoftriloops;
-extern float init;
+extern long double init;
 extern int gail; /* It is either 0 or 1. It is used for grosely asymmetric internal loops */
-extern float prelog;
+extern long double prelog;
 
-extern float tstackm[5][5][6][6];
-extern float tstacke[5][5][6][6];
-extern float tstacki23[5][5][5][5];
+extern long double tstackm[5][5][6][6];
+extern long double tstacke[5][5][6][6];
+extern long double tstacki23[5][5][5][5];
 
 
 #define fourBaseIndex(a, b, c, d) (((a) << 6) + ((b) << 4) + ((c) << 2) + (d))
