+++ gtmfe/include/shapereader.h
@@ -1,14 +1,14 @@
 #ifndef _SHAPEREADER_H_
 #define _SHAPEREADER_H_
 
-extern float* SHAPEarray;
+extern long double* SHAPEarray;
 extern int* SHAPEenergies;
 
 void readSHAPEarray(const char* filename, int seqlength);
 void free_shapeArray(int len);
 void print_shapeArray(int len);
-int calcShapeEnergy(float shapeNumber);
-float shapeModel(float SHAPE_value);
+int calcShapeEnergy(long double shapeNumber);
+long double shapeModel(long double SHAPE_value);
 
 #ifdef __cplusplus
 extern "C"{
